imageProcessing/requirements.txt
flask==3.1.3
pillow==12.1.0
numpy<2.0.0